feat(header): add role-based headers
